matrix: move neighbor-helper functions to matrix/neighbors
